+++ cryoet_data_portal_neuroglancer/state_generator.py
@@ -21,13 +21,42 @@ def _setup_creation(
     url: str | None = None,
     zarr_path: str | None = None,
     scale: tuple[float, float, float] | list[float] | float = 1.0,
+    output_scale: tuple[float, float, float] | list[float] | float | None = None,
 ) -> tuple[str, str, str, str, tuple[float, float, float]]:
+    """Starting point for generating JSON for a layer.
+
+    By taking in some possibly optional parameters for the data setup
+    , we combine these in a way that is useful for the rest of the code
+    around state generation.
+
+    Parameters
+    ----------
+    source: str
+        The source of the data. This can be a file path or a direct URL.
+    name: str | None, optional
+        The name of the layer. If None, the name will be derived from the source.
+    url: str | None, optional
+        The base URL for the data. If None, the source must be a direct URL.
+    zarr_path: str | None, optional
+        The path to the zarr data. If None, the source will be used.
+        This is useful where the source is a URL, not a file path.
+    scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres. Ordering is XYZ.
+        If a single float is provided, it will be used for all three axes.
+        Default is (1.0, 1.0, 1.0).
+
+    Returns
+    -------
+    tuple[str, str, str, str, tuple[float, float, float]]
+        A tuple containing the source, name, url, zarr_path, and scale.
+    """
     name = Path(source).stem if name is None else name
     url = url if url is not None else ""
     zarr_path = zarr_path if zarr_path is not None else source
     scale = get_scale(scale)
+    output_scale = get_scale(output_scale) if output_scale is not None else scale
     source = urljoin(url, source.strip("/")) if url else source
-    return source, name, url, zarr_path, scale
+    return source, name, url, zarr_path, scale, output_scale
 
 
 def _validate_color(color: str | None):
@@ -44,8 +73,53 @@ def generate_point_layer(
     scale: float | tuple[float, float, float] = (1.0, 1.0, 1.0),
     is_visible: bool = True,
     is_instance_segmentation: bool = False,
+    output_scale: float | tuple[float, float, float] | None = None,
 ) -> dict[str, Any]:
-    source, name, url, _, scale = _setup_creation(source, name, url, scale=scale)
+    """Generate neuroglancer JSON state for a single point annotation layer.
+
+    Parameters
+    ----------
+    source: str
+        The source of the data. This can be a file path or a direct URL.
+    name: str | None, optional
+        The name of the layer. If None, the name will be derived from the source.
+    url: str | None, optional
+        The base URL for the data. If None, the source must be a direct URL.
+    color: str, optional
+        The color of the points in hex format. Default is white (#FFFFFF).
+    point_size_multiplier: float, optional
+        The size of the points in the annotation layer. Default is 1.0.
+    scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres. Ordering is XYZ.
+        If a single float is provided, it will be used for all three axes.
+        Default is (1.0, 1.0, 1.0).
+    is_visible: bool, optional
+        Whether the layer should be visible in the viewer on startup. Default is True.
+    is_instance_segmentation: bool, optional
+        An instance segmentation layer is a layer where each point is assigned to a specific instance.
+        This is useful for layers where you want to assign different colors to different instances.
+        Default is False, which means each point is assigned the same color.
+    output_scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres for the output vis.
+        If None, the scale will be used instead.
+
+    Returns
+    -------
+    dict[str, Any]
+        A dictionary containing the JSON state for the layer.
+
+    Raises
+    ------
+    ValueError
+        If the color is not a valid hex string.
+    """
+    source, name, url, _, scale, output_scale = _setup_creation(
+        source,
+        name,
+        url,
+        scale=scale,
+        output_scale=output_scale,
+    )
     _validate_color(color)
     return AnnotationJSONGenerator(
         source=source,
@@ -53,6 +127,7 @@ def generate_point_layer(
         color=color,
         point_size_multiplier=point_size_multiplier,
         scale=scale,
+        output_scale=output_scale,
         is_visible=is_visible,
         is_instance_segmentation=is_instance_segmentation,
     ).to_json()
@@ -68,8 +143,61 @@ def generate_oriented_point_layer(
     scale: float | tuple[float, float, float] = (1.0, 1.0, 1.0),
     is_visible: bool = True,
     is_instance_segmentation: bool = False,
+    is_code_editor_visible: bool = False,
+    output_scale: float | tuple[float, float, float] | None = None,
 ) -> dict[str, Any]:
-    source, name, url, _, scale = _setup_creation(source, name, url, scale=scale)
+    """Generate neuroglancer JSON state for a single oriented point annotation layer.
+
+    Oriented point annotation layers consist of points shown as three direction vectors.
+    The direction vectors are line annotations.
+
+    Parameters
+    ----------
+    source: str
+        The source of the data. This can be a file path or a direct URL.
+    name: str | None, optional
+        The name of the layer. If None, the name will be derived from the source.
+    url: str | None, optional
+        The base URL for the data. If None, the source must be a direct URL.
+    color: str, optional
+        The color of the points in hex format. Default is white (#FFFFFF).
+    point_size_multiplier: float, optional
+        The size of the points in the annotation layer. Default is 1.0.
+    line_width: float, optional
+        The width of the lines in the annotation layer. Default is 1.0.
+    scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres. Ordering is XYZ.
+        If a single float is provided, it will be used for all three axes.
+        Default is (1.0, 1.0, 1.0).
+    is_visible: bool, optional
+        Whether the layer should be visible in the viewer on startup. Default is True.
+    is_instance_segmentation: bool, optional
+        An instance segmentation layer is a layer where each point is assigned to a specific instance.
+        This is useful for layers where you want to assign different colors to different instances.
+        Default is False, which means each point is assigned the same color.
+    is_code_editor_visible: bool, optional
+        Whether the code editor should be visible in the layer settings. Default is False.
+    output_scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres for the output vis.
+        If None, the scale will be used instead.
+
+    Returns
+    -------
+    dict[str, Any]
+        A dictionary containing the JSON state for the layer.
+
+    Raises
+    ------
+    ValueError
+        If the color is not a valid hex string.
+    """
+    source, name, url, _, scale, output_scale = _setup_creation(
+        source,
+        name,
+        url,
+        scale=scale,
+        output_scale=output_scale,
+    )
     _validate_color(color)
     return OrientedPointAnnotationJSONGenerator(
         source=source,
@@ -78,8 +206,10 @@ def generate_oriented_point_layer(
         point_size_multiplier=point_size_multiplier,
         line_width=line_width,
         scale=scale,
+        output_scale=output_scale,
         is_visible=is_visible,
         is_instance_segmentation=is_instance_segmentation,
+        is_code_editor_visible=is_code_editor_visible,
     ).to_json()
 
 
@@ -96,14 +226,75 @@ def generate_segmentation_mask_layer(
     mesh_render_scale: float = 2.0,
     visible_segments: tuple[int, ...] = (1,),
     enable_pick: bool = False,
+    output_scale: float | tuple[float, float, float] | None = None,
 ) -> dict[str, Any]:
-    source, name, url, _, scale = _setup_creation(source, name, url, scale=scale)
+    """Generate neuroglancer JSON state for a single segmentation layer.
+
+    Parameters
+    ----------
+    source: str
+        The source of the data. This can be a file path or a direct URL.
+    name: str | None, optional
+        The name of the layer. If None, the name will be derived from the source.
+    url: str | None, optional
+        The base URL for the data. If None, the source must be a direct URL.
+    color: str, optional
+        The color of the points in hex format. Default is white (#FFFFFF).
+    scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres. Ordering is XYZ.
+        If a single float is provided, it will be used for all three axes.
+        Default is (1.0, 1.0, 1.0).
+    is_visible: bool, optional
+        Whether the layer should be visible in the viewer on startup. Default is True.
+    display_bounding_box: bool, optional
+        Whether to display the bounding box of the layer. Default is False.
+    display_mesh: bool, optional
+        Whether to display the mesh version of the mask. Default is True.
+    highlight_on_hover: bool, optional
+        Whether to highlight the mesh on hover. Default is False.
+    mesh_render_scale: float, optional
+        The scale of the mesh rendering. Default is 2.0.
+        This is the scale of the mesh rendering in the viewer.
+        A lower value will force a higher resolution mesh to be loaded
+        in general.
+    visible_segments: tuple[int, ...], optional
+        The labels segments that should be visible in the viewer. Default is (1,).
+        Many of the segmentation masks have only one segment, which is
+        labeled as 1. So (1,) - the default, shows just that segment.
+    enable_pick: bool, optional
+        Technically, this is the "pick" option in neuroglancer.
+        Having it off also has the effect of making it so that
+        double clicking the segmentation mask does not turn off the
+        segmentation mask in the viewer.
+        Default is False.
+    output_scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres for the output vis.
+        If None, the scale will be used instead.
+
+    Returns
+    -------
+    dict[str, Any]
+        A dictionary containing the JSON state for the layer.
+
+    Raises
+    ------
+    ValueError
+        If the color is not a valid hex string.
+    """
+    source, name, url, _, scale, output_scale = _setup_creation(
+        source,
+        name,
+        url,
+        scale=scale,
+        output_scale=output_scale,
+    )
     _validate_color(color)
     return SegmentationJSONGenerator(
         source=source,
         name=name,
         color=color,
         scale=scale,
+        output_scale=output_scale,
         is_visible=is_visible,
         display_bounding_box=display_bounding_box,
         display_mesh=display_mesh,
@@ -126,14 +317,72 @@ def generate_mesh_layer(
     highlight_on_hover: bool = False,
     mesh_render_scale: float = 2.0,
     visible_segments: tuple[int, ...] = (1,),
+    output_scale: float | tuple[float, float, float] | None = None,
 ) -> dict[str, Any]:
-    source, name, url, _, scale = _setup_creation(source, name, url, scale=scale)
+    """Generate neuroglancer JSON state for a single mesh.
+
+    Techically, this is a "segmentation" layer in neuroglancer.
+    But all the layer shows is a mesh, so we call it a mesh layer.
+
+    Parameters
+    ----------
+    source: str
+        The source of the data. This can be a file path or a direct URL.
+    name: str | None, optional
+        The name of the layer. If None, the name will be derived from the source.
+    url: str | None, optional
+        The base URL for the data. If None, the source must be a direct URL.
+    color: str, optional
+        The color of the points in hex format. Default is white (#FFFFFF).
+    scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres. Ordering is XYZ.
+        If a single float is provided, it will be used for all three axes.
+        Default is (1.0, 1.0, 1.0).
+    is_visible: bool, optional
+        Whether the layer should be visible in the viewer on startup. Default is True.
+    display_bounding_box: bool, optional
+        Whether to display the bounding box of the layer. Default is False.
+    display_mesh: bool, optional
+        Whether to display the mesh version of the mask. Default is True.
+    highlight_on_hover: bool, optional
+        Whether to highlight the mesh on hover. Default is False.
+    mesh_render_scale: float, optional
+        The scale of the mesh rendering. Default is 2.0.
+        This is the scale of the mesh rendering in the viewer.
+        A lower value will force a higher resolution mesh to be loaded
+        in general.
+    visible_segments: tuple[int, ...], optional
+        The labels segments that should be visible in the viewer. Default is (1,).
+        Many of the segmentation masks have only one segment, which is
+        labeled as 1. So (1,) - the default, shows just that segment.
+    output_scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres for the output vis.
+        If None, the scale will be used instead.
+
+    Returns
+    -------
+    dict[str, Any]
+        A dictionary containing the JSON state for the point annotation layer.
+
+    Raises
+    ------
+    ValueError
+        If the color is not a valid hex string.
+    """
+    source, name, url, _, scale, output_scale = _setup_creation(
+        source,
+        name,
+        url,
+        scale=scale,
+        output_scale=output_scale,
+    )
     _validate_color(color)
     return MeshJSONGenerator(
         source=source,
         name=name,
         color=color,
         scale=scale,
+        output_scale=output_scale,
         is_visible=is_visible,
         display_bounding_box=display_bounding_box,
         display_mesh=display_mesh,
@@ -155,13 +404,65 @@ def generate_oriented_point_mesh_layer(
     highlight_on_hover: bool = False,
     mesh_render_scale: float = 4.0,
     visible_segments: tuple[int, ...] = (1,),
+    output_scale: float | tuple[float, float, float] | None = None,
 ) -> dict[str, Any]:
+    """Generate neuroglancer JSON state for a multiple instanced oriented meshes.
+
+    Techically, this is a "segmentation" layer in neuroglancer.
+    But all the layer shows is a mesh, so we call it a mesh layer.
+
+    Parameters
+    ----------
+    source: str
+        The source of the data. This can be a file path or a direct URL.
+    name: str | None, optional
+        The name of the layer. If None, the name will be derived from the source.
+    url: str | None, optional
+        The base URL for the data. If None, the source must be a direct URL.
+    color: str, optional
+        The color of the points in hex format. Default is white (#FFFFFF).
+    scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres. Ordering is XYZ.
+        If a single float is provided, it will be used for all three axes.
+        Default is (1.0, 1.0, 1.0).
+    is_visible: bool, optional
+        Whether the layer should be visible in the viewer on startup. Default is True.
+    display_bounding_box: bool, optional
+        Whether to display the bounding box of the layer. Default is False.
+    display_mesh: bool, optional
+        Whether to display the mesh version of the mask. Default is True.
+    highlight_on_hover: bool, optional
+        Whether to highlight the mesh on hover. Default is False.
+    mesh_render_scale: float, optional
+        The scale of the mesh rendering. Default is 2.0.
+        This is the scale of the mesh rendering in the viewer.
+        A lower value will force a higher resolution mesh to be loaded
+        in general.
+    visible_segments: tuple[int, ...], optional
+        The labels segments that should be visible in the viewer. Default is (1,).
+        Many of the segmentation masks have only one segment, which is
+        labeled as 1. So (1,) - the default, shows just that segment.
+    output_scale: float | tuple[float, float, float], optional
+        The scale/resolution of the data in metres for the output vis.
+        If None, the scale will be used instead.
+
+    Returns
+    -------
+    dict[str, Any]
+        A dictionary containing the JSON state for the layer.
+
+    Raises
+    ------
+    ValueError
+        If the color is not a valid hex string.
+    """
     return generate_mesh_layer(
         source=source,
         name=name,
         url=url,
         color=color,
         scale=scale,
+        output_scale=output_scale,
         is_visible=is_visible,
         display_bounding_box=display_bounding_box,
         display_mesh=display_mesh,
@@ -181,7 +482,7 @@ def generate_image_layer(
     mean: float | None = None,
     rms: float | None = None,
     is_visible: bool = True,
-    has_volume_rendering_shader: bool = False,
+    has_volume_rendering_shader: bool | None = None,
     twodee_contrast_limits: tuple[float, float] | None = None,
     threedee_contrast_limits: tuple[float, float] | None = None,
     volume_rendering_is_visible: bool = False,
@@ -189,21 +490,104 @@ def generate_image_layer(
     can_hide_high_values_in_neuroglancer: bool | None = None,
     blend: str = "additive",
     opacity: float = 1.0,
+    is_code_editor_visible: bool = False,
+    output_scale: float | tuple[float, float, float] | None = None,
 ) -> dict[str, Any]:
-    """Generates JSON for an image layer with optional contrast limits.
+    """Generate neuroglancer JSON state for a tomogram (image) layer.
 
-    Note, if twodee_contrast_limits are not provided, the contrast limits will be calculated using the mean and rms values. If threedee_contrast_limits are not provided, the contrast limits will be the same as the twodee_contrast_limits.
+    Parameters
+    ----------
+    source: str
+        The source of the data. This can be a file path or a direct URL.
+    scale: float | tuple[float, float, float]
+        The scale/resolution of the data in metres. Ordering is XYZ.
+    size: dict[str, float]
+        The size of the data in each dimension in in pixels.
+        (e.g. {"x": 512, "y": 400, "z": 256}).
+        The keys are "x", "y", and "z".
+        This is used to set the center position of the data in the viewer
+        and the zoom level in 2D and 3D views.
+    name: str | None, optional
+        The name of the layer. If None, the name will be derived from the source.
+    url: str | None, optional
+        The base URL for the data. If None, the source must be a direct URL.
+    start: dict[str, float] | None, optional
+        The starting position of the data, usually not needed and can be left as None.
+        Default is None, which is essentially the same as {"x": 0, "y": 0, "z": 0}.
+        The keys are "x", "y", and "z".
+    mean: float | None, optional
+        The mean value of the data. Default is None.
+    rms: float | None, optional
+        The root mean square value of the data. Default is None.
+        If contrast limits are not provided, at least one of the mean or rms
+        should be provided.
+    is_visible: bool, optional
+        Whether the layer should be visible in the viewer on startup. Default is True.
+    has_volume_rendering_shader: bool | None, optional
+        Whether to create the layer with a volume rendering shader.
+        Default is None, which means it will be set to True if
+        volume_rendering_is_visible is True or threedee_contrast_limits
+        is not None.
+    twodee_contrast_limits: tuple[float, float] | None, optional
+        The contrast limits for the 2D view. Default is None.
+        If None, the mean and rms values will be used to set the contrast limits.
+        If both mean and rms are None, the contrast limits will be set to
+        (0, 1) by default.
+    threedee_contrast_limits: tuple[float, float] | None, optional
+        The contrast limits for the 3D view. Default is None.
+        If None, these default to the same as the 2D contrast limits.
+    volume_rendering_is_visible: bool, optional
+        Whether to show the volume rendering in the viewer. Default is False.
+    volume_rendering_gain: float, optional
+        The gain for the volume rendering. Default is -7.8.
+        This is an exponential gain value that is used to adjust the alpha
+        of the volume rendering. A lower value will make the volume rendering
+        darker / more transparent, while a higher value will make it brighter and
+        more opaque.
+    can_hide_high_values_in_neuroglancer: bool | None, optional
+        Whether to allow the user to hide high values in the neuroglancer viewer.
+        High values are defined as values above the contrast limits.
+        Having this set to True spawns a control in the layer side panel
+        that allows the user to hide high values for this layer.
+        Default is None, which means it will be set to True if
+        has_volume_rendering_shader is True.
+    blend: str, optional
+        The blending mode for the layer. Default is "additive".
+        The only other option is "default"
+    opacity: float, optional
+        The opacity of the layer. Default is 1.0.
+        This is the opacity of the layer in the viewer.
+        A lower value will make the layer more transparent, while a higher
+        value will make it more opaque.
+    is_code_editor_visible: bool, optional
+        Whether the code editor should be visible in the layer settings.
+        Default is False.
+    output_scale: float | tuple[float, float, float] | None, optional
+        The scale/resolution of the data in metres for the output vis.
+        If None, the scale will be used instead.
+
+    Returns
+    -------
+    dict[str, Any]
+        A dictionary containing the JSON state for the layer.
     """
     # If volume rendering is visible, set the flag to True for the relevant shader
-    if not has_volume_rendering_shader and volume_rendering_is_visible:
-        has_volume_rendering_shader = True
-    source, name, url, _, scale = _setup_creation(source, name, url, scale=scale)
+    if has_volume_rendering_shader is None:
+        has_volume_rendering_shader = volume_rendering_is_visible or threedee_contrast_limits is not None
+    source, name, url, _, scale, output_scale = _setup_creation(
+        source,
+        name,
+        url,
+        scale=scale,
+        output_scale=output_scale,
+    )
     if can_hide_high_values_in_neuroglancer is None:
         can_hide_high_values_in_neuroglancer = has_volume_rendering_shader
     return ImageJSONGenerator(
         source=source,
         name=name,
         scale=scale,
+        output_scale=output_scale,
         size=size,
         start=start,
         mean=mean,
@@ -217,6 +601,7 @@ def generate_image_layer(
         can_hide_high_values_in_neuroglancer=can_hide_high_values_in_neuroglancer,
         blend=blend,
         opacity=opacity,
+        is_code_editor_visible=is_code_editor_visible,
     ).to_json()
 
 
@@ -231,6 +616,11 @@ def generate_image_volume_layer(
     blend: str = "additive",
     opacity: float = 1.0,
 ) -> dict[str, Any]:
+    """
+    ! This function is deprecated, please use "generate_segmentation_mask_layer" instead !
+
+    It was originally used to visualize segmentations via volume rendering.
+    """
     source, name, url, _, scale = _setup_creation(source, name, url, scale=scale)
     _validate_color(color)
     return ImageVolumeJSONGenerator(
@@ -252,8 +642,53 @@ def combine_json_layers(
     projection_quaternion: list[float] | None = None,
     set_slices_visible_in_3d: bool | None = None,
     show_axis_lines: bool = True,
+    use_old_neuroglancer_layout: bool = True,
+    gpu_memory_limit_gb: float = 1.5,
+    voxel_size_scale: float = 1.0,
 ) -> dict[str, Any]:
-    """Note, if set_slices_visible_in_3d is not provided, it will be set to False if there are any image layers in the list with volume rendering."""
+    """Take a list of JSON layers and combine them into a neuroglancer state.
+
+    Parameters
+    ----------
+    layers: list[dict[str, Any]]
+        The list of JSON state neuroglancer layers to combine.
+    scale: float | tuple[float, float, float] | list[float]
+        The scale/resolution for the global neuroglancer co-ordinate space in "units". Ordering is XYZ.
+    units: str, optional
+        The units for the scale. Default is "m".
+        For example, "nm" for nanometres, "um" for micrometres, etc.
+    projection_quaternion: list[float] | None, optional
+        The quaternion for the projection orientation. Default is None.
+        If None, the default is a rotation of 45 degrees around the x-axis.
+        Easiest is to use the scipy.spatial.transform.Rotation class to generate
+        the quaternion. For example:
+        >>> from scipy.spatial.transform import Rotation
+        >>> projection_quaternion = Rotation.from_euler(seq="xyz", angles=(45, 0, 0), degrees=True).as_quat()
+    set_slices_visible_in_3d: bool | None, optional
+        Whether to show the slices in 3D. Default is None.
+        If None, it will be set to False if there are any image layers in the list with volume rendering.
+    show_axis_lines: bool, optional
+        Whether to show the axis lines in the 3D view. Default is True.
+    enable_layer_color_legend: bool, optional
+        Whether to enable the layer color legend in the 3D view. Default is True.
+    use_old_neuroglancer_layout: bool, optional
+        Whether to use the old neuroglancer panels layout. Default is True.
+    gpu_memory_limit_gb: float, optional
+        The GPU memory limit for the neuroglancer viewer in GB. Default is 1.5.
+        Going above 2GB would not be recommended as it may cause the browser to crash.
+        In between 1.5 and 2GB could be experimented with.
+        The default neuroglancer uses is actually 1GB GPU memory, and 2GB system memory.
+
+    Returns
+    -------
+    dict[str, Any]
+        A dictionary containing the combined JSON state for the layers.
+
+    Raises
+    ------
+    ValueError
+        If the scale is not a valid tuple or list of floats.
+    """
     image_layers = [layer for layer in layers if layer["type"] == "image"]
     if set_slices_visible_in_3d is None:
         set_slices_visible_in_3d = not any(layer["volumeRendering"] == "on" for layer in image_layers)
@@ -261,6 +696,8 @@ def combine_json_layers(
     scale = get_scale(scale)
     if projection_quaternion is None:
         projection_quaternion = Rotation.from_euler(seq="xyz", angles=(45, 0, 0), degrees=True).as_quat()
+
+    layout = "4panel" if use_old_neuroglancer_layout else "4panel-alt"
     combined_json = {
         "dimensions": {dim: [res, units] for dim, res in zip("xyz", scale, strict=False)},
         "crossSectionScale": 1.8,
@@ -272,7 +709,8 @@ def combine_json_layers(
             "side": "left",
         },
         "crossSectionBackgroundColor": "#000000",
-        "layout": "4panel",
+        "hideCrossSectionBackground3D": True,
+        "layout": layout,
         "showSlices": set_slices_visible_in_3d,
         "showAxisLines": show_axis_lines,
         "layerListPanel": {
@@ -301,10 +739,11 @@ def combine_json_layers(
                 "verticalStacking": False,
             },
         },
+        "gpuMemoryLimit": int(gpu_memory_limit_gb * 1e9),
     }
     if len(image_layers) > 0 and "_position" in image_layers[0]:
-        combined_json["position"] = image_layers[0]["_position"]
-        combined_json["crossSectionScale"] = image_layers[0]["_crossSectionScale"]
-        combined_json["projectionScale"] = image_layers[0]["_projectionScale"]
+        combined_json["position"] = [x * voxel_size_scale for x in image_layers[0]["_position"]]
+        combined_json["crossSectionScale"] = voxel_size_scale * image_layers[0]["_crossSectionScale"]
+        combined_json["projectionScale"] = voxel_size_scale * image_layers[0]["_projectionScale"]
 
     return combined_json